scraper_service/requirements.txt
fastapi
uvicorn
playwright
supabase==2.7.4
python-dotenv
requests
apscheduler
websockets==13.1
httpx==0.27.0